Move inventory scan to its own module
Organize the scanning code and make it more clear what's happening.
